prod-requirements.txt
pandas
nltk
numpy
fastapi[all]
watchdog
sentence-transformers 
faiss-cpu